pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.68.0",
    "black>=25.9.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]
